Add setting for JSON indentation
Closes #1976
