+++ src/Spectre.Console.Tests/Unit/Widgets/JsonTextTests.cs
@@ -19,4 +19,25 @@ public sealed class JsonTextTests
         // Then
         return Verifier.Verify(console.Output);
     }
+
+    [Fact]
+    [Expectation("Render_Compact_Json")]
+    public Task Should_Render_Compact_Json()
+    {
+        // Given
+        var console = new TestConsole().Size(new Size(80, 15));
+        var json = EmbeddedResourceReader
+            .LoadResourceStream("Spectre.Console.Tests/Data/example.json")
+            .ReadText();
+
+        // When
+        console.Write(
+            new Panel(
+                    new JsonText(json)
+                        .Indentation(" "))
+                .Header("Some JSON"));
+
+        // Then
+        return Verifier.Verify(console.Output);
+    }
 }
